fix: events not-found page for both localized paths
